+++ app/src/main/java/me/saket/dank/urlparser/ImgurAlbumLink.java
@@ -7,12 +7,14 @@ import androidx.annotation.Nullable;
 import com.google.auto.value.AutoValue;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
 @AutoValue
-public abstract class ImgurAlbumLink extends MediaLink implements Parcelable {
+public abstract class ImgurAlbumLink extends MediaAlbumLink<ImgurLink> implements Parcelable {
 
+  @NotNull
   @Override
   public abstract String unparsedUrl();
 
@@ -21,31 +23,14 @@ public abstract class ImgurAlbumLink extends MediaLink implements Parcelable {
   public abstract String albumUrl();
 
   @Nullable
+  @Override
   public abstract String albumTitle();
 
-  public boolean hasAlbumTitle() {
-    //noinspection ConstantConditions
-    return albumTitle() != null && !albumTitle().isEmpty();
-  }
-
+  @Override
   public abstract String coverImageUrl();
 
+  @Override
   public abstract List<ImgurLink> images();
-
-  @Override
-  public Link.Type type() {
-    return Link.Type.MEDIA_ALBUM;
-  }
-
-  @Override
-  public String highQualityUrl() {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public String lowQualityUrl() {
-    throw new UnsupportedOperationException();
-  }
 
   @Override
   public String cacheKey() {