+++ app/src/main/java/me/saket/dank/urlparser/RedditSubredditLink.java
@@ -3,10 +3,12 @@ package me.saket.dank.urlparser;
 import android.os.Parcelable;
 
 import com.google.auto.value.AutoValue;
+import org.jetbrains.annotations.NotNull;
 
 @AutoValue
 public abstract class RedditSubredditLink extends RedditLink implements Parcelable {
 
+  @NotNull
   @Override
   public abstract String unparsedUrl();