+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+-   Support for reddit galleries ([#312](https://github.com/Tunous/Dawn/pull/312))
+
 ## [0.10.0] - 2020-08-14
 
 ### Added